refactoring: entities refactoring
