+++ src/main/java/org/itmo/eventapp/main/model/dto/request/TaskFilterRequest.java
@@ -1,6 +1,6 @@
 package org.itmo.eventapp.main.model.dto.request;
 
-import org.itmo.eventapp.main.model.entity.TaskStatus;
+import org.itmo.eventapp.main.model.entity.enums.TaskStatus;
 
 import java.time.LocalDateTime;
