+++ src/main/java/org/itmo/eventapp/main/model/dto/response/TaskResponse.java
@@ -1,6 +1,6 @@
 package org.itmo.eventapp.main.model.dto.response;
 
-import org.itmo.eventapp.main.model.entity.TaskStatus;
+import org.itmo.eventapp.main.model.entity.enums.TaskStatus;
 
 import java.time.LocalDateTime;
