+++ src/main/java/org/itmo/eventapp/main/model/entity/Event.java
@@ -2,19 +2,19 @@ package org.itmo.eventapp.main.model.entity;
 
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.hibernate.annotations.JdbcType;
-import org.hibernate.annotations.Type;
 import org.hibernate.dialect.PostgreSQLEnumJdbcType;
+import org.itmo.eventapp.main.model.entity.enums.EventFormat;
+import org.itmo.eventapp.main.model.entity.enums.EventStatus;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 @Entity
-@Table(name = "event")
-@Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
 public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,10 +31,8 @@ public class Event {
 
     private String title;
 
-    @Column(name = "short_description")
     private String shortDescription;
 
-    @Column(name = "full_description")
     private String fullDescription;
 
     @Enumerated(EnumType.STRING)
@@ -45,66 +43,21 @@ public class Event {
     @JdbcType(PostgreSQLEnumJdbcType.class)
     private EventStatus status;
 
-    @Column(name = "registration_start")
     private LocalDateTime registrationStart;
 
-    @Column(name = "registration_end")
     private LocalDateTime registrationEnd;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "parent_id")
     private Event parent;
 
-    @Column(name = "participant_limit")
     private int participantLimit;
 
-    @Column(name = "participant_age_lowest")
-    private int participantsAgeLowest;
+    private int participantAgeLowest;
 
-    @Column(name = "participant_age_highest")
-    private int participantsAgeHighest;
+    private int participantAgeHighest;
 
-    @Column(name = "preparing_start")
     private LocalDateTime preparingStart;
 
-    @Column(name = "preparing_end")
     private LocalDateTime preparingEnd;
-
-    public Event() {
-    }
-
-    public Event(
-            Place place,
-            LocalDateTime start,
-            LocalDateTime end,
-            String title,
-            String shortDescription,
-            String fullDescription,
-            EventFormat format,
-            EventStatus status,
-            LocalDateTime registrationStart,
-            LocalDateTime registrationEnd,
-            Event parent,
-            int participantLimit,
-            int participantsAgeLowest,
-            int participantsAgeHighest,
-            LocalDateTime preparingStart,
-            LocalDateTime preparingEnd) {
-        this.place = place;
-        this.start = start;
-        this.end = end;
-        this.title = title;
-        this.shortDescription = shortDescription;
-        this.fullDescription = fullDescription;
-        this.format = format;
-        this.status = status;
-        this.registrationStart = registrationStart;
-        this.registrationEnd = registrationEnd;
-        this.parent = parent;
-        this.participantLimit = participantLimit;
-        this.participantsAgeLowest = participantsAgeLowest;
-        this.participantsAgeHighest = participantsAgeHighest;
-        this.preparingStart = preparingStart;
-        this.preparingEnd = preparingEnd;
-    }
 }
