+++ src/main/java/org/itmo/eventapp/main/model/entity/Participant.java
@@ -1,13 +1,13 @@
 package org.itmo.eventapp.main.model.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 @Entity
-@Table(name = "participant")
-@Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
 public class Participant {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -17,7 +17,6 @@ public class Participant {
 
     private String email;
 
-    @Column(name = "additional_info")
     private String additionalInfo;
 
     private boolean visited;
@@ -25,15 +24,4 @@ public class Participant {
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "event_id")
     private Event event;
-
-    public Participant() {
-    }
-
-    public Participant(String name, String email, String additionalInfo, boolean visited, Event event) {
-        this.name = name;
-        this.email = email;
-        this.additionalInfo = additionalInfo;
-        this.visited = visited;
-        this.event = event;
-    }
 }
