+++ src/main/java/org/itmo/eventapp/main/model/entity/Privilege.java
@@ -1,16 +1,16 @@
 package org.itmo.eventapp.main.model.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.hibernate.annotations.JdbcType;
-import org.hibernate.annotations.Type;
 import org.hibernate.dialect.PostgreSQLEnumJdbcType;
+import org.itmo.eventapp.main.model.entity.enums.PrivilegeType;
 
 @Entity
-@Table(name = "privilege")
-@Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
 public class Privilege {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -23,13 +23,4 @@ public class Privilege {
     @Enumerated(EnumType.STRING)
     @JdbcType(PostgreSQLEnumJdbcType.class)
     private PrivilegeType type;
-
-    public Privilege() {
-    }
-
-    public Privilege(String name, String description, PrivilegeType type) {
-        this.name = name;
-        this.description = description;
-        this.type = type;
-    }
 }
