+++ src/main/java/org/itmo/eventapp/main/model/entity/Role.java
@@ -1,19 +1,19 @@
 package org.itmo.eventapp.main.model.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.hibernate.annotations.JdbcType;
-import org.hibernate.annotations.Type;
 import org.hibernate.dialect.PostgreSQLEnumJdbcType;
+import org.itmo.eventapp.main.model.entity.enums.RoleType;
 
 import java.util.HashSet;
 import java.util.Set;
 
 @Entity
-@Table(name = "role")
-@Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -28,17 +28,8 @@ public class Role {
     private RoleType type;
 
     @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    @JoinTable(name="role_privilege",
-            joinColumns=  @JoinColumn(name="role_id", referencedColumnName="id"),
-            inverseJoinColumns= @JoinColumn(name="privilege_id", referencedColumnName="id") )
-    private Set<Privilege> privileges = new HashSet<Privilege>();
-
-    public Role() {
-    }
-
-    public Role(String name, String description, RoleType type) {
-        this.name = name;
-        this.description = description;
-        this.type = type;
-    }
+    @JoinTable(name = "role_privilege",
+            joinColumns = @JoinColumn(name = "role_id", referencedColumnName = "id"),
+            inverseJoinColumns = @JoinColumn(name = "privilege_id", referencedColumnName = "id"))
+    private Set<Privilege> privileges = new HashSet<>();
 }
