--- src/main/java/org/itmo/eventapp/main/model/entity/RoleType.java
@@ -1,6 +1,0 @@
-package org.itmo.eventapp.main.model.entity;
-
-public enum RoleType {
-    SYSTEM,
-    EVENT,
-}
