+++ src/main/java/org/itmo/eventapp/main/model/entity/Task.java
@@ -1,19 +1,18 @@
 package org.itmo.eventapp.main.model.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.hibernate.annotations.JdbcType;
-import org.hibernate.annotations.Type;
 import org.hibernate.dialect.PostgreSQLEnumJdbcType;
+import org.itmo.eventapp.main.model.entity.enums.TaskStatus;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 @Entity
-@Table(name = "task")
-@Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
 public class Task {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -37,8 +36,7 @@ public class Task {
 
     @Enumerated(EnumType.STRING)
     @JdbcType(PostgreSQLEnumJdbcType.class)
-    @Column(name = "status")
-    private TaskStatus taskStatus;
+    private TaskStatus status;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "place_id")
@@ -46,30 +44,5 @@ public class Task {
 
     private LocalDateTime deadline;
 
-    @Column(name = "notification_deadline")
     private LocalDateTime notificationDeadline;
-
-    public Task() {
-    }
-
-    public Task(
-            Event event,
-            User assignee,
-            User assigner,
-            String title,
-            String description,
-            TaskStatus taskStatus,
-            Place place,
-            LocalDateTime deadline,
-            LocalDateTime notificationDeadline) {
-        this.event = event;
-        this.assignee = assignee;
-        this.assigner = assigner;
-        this.title = title;
-        this.description = description;
-        this.taskStatus = taskStatus;
-        this.place = place;
-        this.deadline = deadline;
-        this.notificationDeadline = notificationDeadline;
-    }
 }
