+++ src/main/java/org/itmo/eventapp/main/model/entity/UserLoginInfo.java
@@ -1,19 +1,18 @@
 package org.itmo.eventapp.main.model.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.hibernate.annotations.JdbcType;
-import org.hibernate.annotations.Type;
 import org.hibernate.dialect.PostgreSQLEnumJdbcType;
+import org.itmo.eventapp.main.model.entity.enums.EmailStatus;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 @Entity
-@Table(name = "user_login_info")
-@Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
 public class UserLoginInfo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -27,32 +26,15 @@ public class UserLoginInfo {
 
     @Enumerated(EnumType.STRING)
     @JdbcType(PostgreSQLEnumJdbcType.class)
-    @Column(name = "email_status")
     private EmailStatus emailStatus;
 
-    @Column(name = "password_hash")
     private String passwordHash;
 
-    @Column(name = "resettoken")
     private String resetToken;
 
     @OneToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "registration_id")
     private RegistrationRequest registration;
 
-    @Column(name = "last_login_date")
     private LocalDateTime lastLoginDate;
-
-    public UserLoginInfo() {
-    }
-
-    public UserLoginInfo(User user, String email, EmailStatus emailStatus, String passwordHash, String resetToken, RegistrationRequest registration, LocalDateTime lastLoginDate) {
-        this.user = user;
-        this.email = email;
-        this.emailStatus = emailStatus;
-        this.passwordHash = passwordHash;
-        this.resetToken = resetToken;
-        this.registration = registration;
-        this.lastLoginDate = lastLoginDate;
-    }
 }
