+++ src/main/java/org/itmo/eventapp/main/model/entity/UserNotificationInfo.java
@@ -1,13 +1,13 @@
 package org.itmo.eventapp.main.model.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 @Entity
-@Table(name = "user_notification_info")
-@Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
 public class UserNotificationInfo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -18,19 +18,7 @@ public class UserNotificationInfo {
 
     private String surname;
 
-    @Column(name = "enable_push_notifications")
     private String enablePushNotifications;
 
-    @Column(name = "enable_email_notifications")
     private String enableEmailNotifications;
-
-    public UserNotificationInfo() {
-    }
-
-    public UserNotificationInfo(String[] devices, String surname, String enablePushNotifications, String enableEmailNotifications) {
-        this.devices = devices;
-        this.surname = surname;
-        this.enablePushNotifications = enablePushNotifications;
-        this.enableEmailNotifications = enableEmailNotifications;
-    }
 }
