+++ src/main/java/org/itmo/eventapp/main/model/entity/enums/RoleType.java
@@ -1,0 +1,6 @@
+package org.itmo.eventapp.main.model.entity.enums;
+
+public enum RoleType {
+    SYSTEM,
+    EVENT,
+}
